requirements.txt
IPy==1.0
chardet==3.0.4
idna==2.10
requests==2.24.0
urllib3==1.25.11
tqdm>=4.51.0
baidu-aip==2.2.18.0
playsound==1.2.2
